@@ -18,33 +18,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.5",
-    "@angular/cdk": "^13.3.5",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
+    "@angular/animations": "14.2.5",
+    "@angular/cdk": "14.2.4",
+    "@angular/common": "14.2.5",
+    "@angular/compiler": "14.2.5",
+    "@angular/core": "14.2.5",
     "@angular/fire": "^7.4.1",
-    "@angular/forms": "~13.3.0",
+    "@angular/forms": "14.2.5",
     "@angular/google-maps": "^13.3.8",
-    "@angular/localize": "^14.0.0",
-    "@angular/material": "^13.3.5",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/localize": "14.2.5",
+    "@angular/material": "14.2.4",
+    "@angular/platform-browser": "14.2.5",
+    "@angular/platform-browser-dynamic": "14.2.5",
+    "@angular/router": "14.2.5",
     "@coreui/angular": "4",
     "@firebase/storage": "^0.9.9",
     "@google/maps": "^1.1.3",
     "@googlemaps/js-api-loader": "^1.14.3",
     "@googlemaps/markerclusterer": "^2.0.8",
     "@material/theme": "^14.0.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
+    "@nestjs/common": "9.0.0",
+    "@nestjs/core": "9.0.0",
     "@nestjs/cqrs": "^8.0.4",
     "@nestjs/jwt": "^9.0.0",
     "@nestjs/passport": "^9.0.0",
-    "@nestjs/platform-express": "^8.0.0",
+    "@nestjs/platform-express": "9.0.0",
     "@ng-bootstrap/ng-bootstrap": "^12.1.2",
-    "@nrwl/angular": "14.0.5",
+    "@nrwl/angular": "14.8.4",
     "@popperjs/core": "^2.10.2",
     "@prisma/client": "^3.13.0",
     "@swimlane/ngx-charts": "^20.1.0",
@@ -71,46 +71,47 @@
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.4.0",
+    "rxjs": "~7.5.0",
     "seedrandom": "^3.0.5",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.3",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "14.0.5",
-    "@nrwl/cypress": "14.0.5",
-    "@nrwl/eslint-plugin-nx": "14.0.5",
-    "@nrwl/jest": "14.0.5",
-    "@nrwl/js": "14.0.5",
-    "@nrwl/linter": "14.0.5",
-    "@nrwl/nest": "14.0.5",
-    "@nrwl/node": "14.0.5",
-    "@nrwl/nx-cloud": "^14.0.3",
-    "@nrwl/workspace": "14.0.5",
-    "@types/jest": "27.4.1",
-    "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
+    "@angular-devkit/build-angular": "14.2.5",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "~14.2.0",
+    "@angular/compiler-cli": "14.2.5",
+    "@angular/language-service": "14.2.5",
+    "@nestjs/schematics": "9.0.0",
+    "@nestjs/testing": "9.0.0",
+    "@nrwl/cli": "14.8.4",
+    "@nrwl/cypress": "14.8.4",
+    "@nrwl/eslint-plugin-nx": "14.8.4",
+    "@nrwl/jest": "14.8.4",
+    "@nrwl/js": "14.8.4",
+    "@nrwl/linter": "14.8.4",
+    "@nrwl/nest": "14.8.4",
+    "@nrwl/node": "14.8.4",
+    "@nrwl/nx-cloud": "14.7.0",
+    "@nrwl/workspace": "14.8.4",
+    "@types/jest": "28.1.8",
+    "@types/node": "18.7.1",
+    "@typescript-eslint/eslint-plugin": "5.36.1",
+    "@typescript-eslint/parser": "5.36.1",
     "cypress": "9.7.0",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "nx": "14.0.5",
-    "prettier": "^2.5.1",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.2",
+    "nx": "14.8.4",
+    "prettier": "2.6.2",
     "prisma": "^4.1.0",
-    "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
-    "typescript": "~4.6.2"
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4"
   }
 }
